Fix condvar_wait_until abstime type (u64→usize) + FreeRTOS investigation
- Fix shim _z_condvar_wait_until: abstime is *const usize (z_clock_t)
  not *const u64 — reading 8 bytes from 4-byte value caused garbage timeout
- FreeRTOS talker works correctly when run manually with zenohd
- All 12 platform unit tests pass (clock, memory, sleep, random, mutex,
  recursive mutex, task+join, condvar, condvar timeout, concurrent tasks,
  mutex contention, background task contention)
- E2E test still fails — likely timing issue when listener + talker QEMU
  instances run simultaneously; needs further investigation

+++ packages/zpico/zpico-platform-shim/src/shim.rs
@@ -358,9 +358,9 @@ pub extern "C" fn _z_condvar_wait(cv: *mut ZCondvar, m: *mut ZMutex) -> i8 {
 pub extern "C" fn _z_condvar_wait_until(
     cv: *mut ZCondvar,
     m: *mut ZMutex,
-    abstime: *const u64,
+    abstime: *const usize, // z_clock_t = TickType_t (u32) on FreeRTOS, void* on bare-metal
 ) -> i8 {
-    let t = unsafe { *abstime };
+    let t = unsafe { *abstime } as u64;
     P::condvar_wait_until(cv as *mut c_void, m as *mut c_void, t)
 }
